Library_management_1.0

+++ library_management/src/books/views.py
@@ -1,12 +1,11 @@
-import jwt
 from flask import make_response, jsonify, request
-from flask_jwt_extended import jwt_required, current_user, decode_token, get_jwt_identity
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from flask_restful import Resource
 from .models import User, Books, BooksStatus
 from .schemas import UserSchema, BooksSchema, BooksStatusSchema
 from ..utils import api, db
 from ..utils import security
-from sqlalchemy import and_, func, desc, or_, asc
+from sqlalchemy import and_, desc, or_, asc
 
 
 
@@ -16,16 +15,23 @@ class UserResource(Resource):
 
     @jwt_required
     def get(self, slug):
+        current_user = get_jwt_identity()
 
-        print(current_user['is_admin'])
-        user = self.model.query.filter(User.id == slug).first()
-        if user and user.is_admin and user.active:
-            users = User.query.all()
-            print(users)
-            data = self.schema().dump(users, many=True)
-            print(data)
+        if current_user['is_admin'] and current_user['active']:
+            users = User.query
+            if '__page' in request.args and request.args.get('__page') and '__limit' in request.args and request.args.get('__limit'):
+                page = request.args.get('__page')
+                limit = request.args.get('__limit')
+                users = users.limit(limit).offset(int(limit) * int(page)-1)
+            if '__user_name__equal' in request.args and request.args.get('__user_name__equal'):
+                user_name =request.args.get('__user_name__equal')
+                data = self.schema().dump(users.filter(User.name == user_name).all(), many=True)
+                return make_response(jsonify({'User_by_name': data}))
+            data = self.schema().dump(users.order_by(asc(User.name)).all(), many=True)
+
             return make_response(jsonify({'Users': data}))
-        if user and user is not user.is_admin and user.active:
+        if not current_user['is_admin'] and current_user['active']:
+            user = self.model.query.filter(User.id == slug).first()
             data = self.schema().dump(user)
             return make_response(jsonify({'Users': data}),200)
 
@@ -45,6 +51,8 @@ class UserResource(Resource):
             UserSchema().load(data, session=db.session, instance=obj)
             db.session.commit()
             return 'User Updated Successfully',200
+        else:
+            return 'You have Permission to do admin work', 200
 
     @jwt_required
     def delete(self, slug):
@@ -54,11 +62,13 @@ class UserResource(Resource):
             obj = User.query.filter(User.id == slug).first()
         if not current_user['is_admin']:
             obj = User.query.filter(or_(User.id == current_user['id'] ), (User.id == slug), (current_user['id'] == slug)).first()
+            if not obj:
+                return 'You have Permission to do admin work', 200
         if not obj:
             return make_response(jsonify({'error': 'Resource not found'}), 404)
         db.session.delete(obj)
         db.session.commit()
-        return make_response(jsonify({}), 204)
+        return 'User Deleted Successfully',204
 
 
 api.add_resource(UserResource, '/user/<slug>', endpoint='user')
@@ -69,41 +79,58 @@ class BooksResource(Resource):
     model = Books
     schema = BooksSchema
 
-    def get(self, slug=None):
-        if slug:
-            books = Books.query.get(slug)
-            type = False
-        else:
-            books = Books.query.all()
-            type = True
+    @jwt_required
+    def get(self):
+        books = Books.query
+        if '__page' in request.args and request.args.get('__page') and '__limit' in request.args and request.args.get(
+                '__limit'):
+            page = request.args.get('__page')
+            limit = request.args.get('__limit')
+            books = books.limit(limit).offset(int(limit) * int(page) - 1)
+        if '__book_name__equal' in request.args and request.args.get('__book_name__equal'):
+            book_name = request.args.get('__book_name__equal')
+            data = self.schema().dump(Books.filter(Books.name == book_name).all(), many=True)
+            return make_response(jsonify({"book_by_name": data}))
+        data = BooksSchema().dump(books.order_by(asc(Books.author)).all(), many=True)
 
-        data = BooksSchema().dump(books, many=type)
-        print(data)
+        return make_response(jsonify({"books": data}))
 
-        return make_response(jsonify({"book": data}))
-
+    @jwt_required
     def post(self):
-        data = request.json
-        obj = BooksSchema().load(data, session=db.session, many=False)
-        db.session.add(obj)
-        db.session.commit()
-        return 'User Added Successfully',201
+        current_user = get_jwt_identity()
+        if current_user['is_admin']:
+            data = request.json
+            obj = BooksSchema().load(data, session=db.session, many=False)
+            db.session.add(obj)
+            db.session.commit()
+            return 'User Added Successfully',201
+        else:
+            return 'You have Permission to do admin work', 200
 
     def patch(self, slug):
-        obj = Books.query.get(slug)
-        data = request.json
-        BooksSchema().load(data, session=db.session, instance=obj)
+        current_user = get_jwt_identity()
+        if current_user['is_admin']:
+            obj = Books.query.get(slug)
+            data = request.json
+            BooksSchema().load(data, session=db.session, instance=obj)
 
-        db.session.commit()
-        return 'User Updated Successfully',200
+            db.session.commit()
+            return 'User Updated Successfully',200
+        else:
+            return 'You have Permission to do admin work', 200
+
 
     def delete(self, slug):
-        obj = self.model.query.get(slug)
-        if not obj:
-            return make_response(jsonify({'error': 'Resource not found'}), 404)
-        db.session.delete(obj)
-        db.session.commit()
-        return make_response(jsonify('success'), 204)
+        current_user = get_jwt_identity()
+        if current_user['is_admin']:
+            obj = self.model.query.get(slug)
+            if not obj:
+                return make_response(jsonify({'error': 'Resource not found'}), 404)
+            db.session.delete(obj)
+            db.session.commit()
+            return make_response(jsonify('success'), 204)
+        else:
+            return 'You have Permission to do admin work', 200
 
 
 api.add_resource(BooksResource, '/books/<slug>', endpoint='books')
@@ -114,13 +141,20 @@ class BooksStatusResource(Resource):
     model = BooksStatus
     schema = BooksStatusSchema
 
+    @jwt_required
     def get(self):
-        books_status = BooksStatus.query.all()
+        books_status = BooksStatus.query
+        if '__page' in request.args and request.args.get('__page') and '__limit' in request.args and request.args.get(
+                '__limit'):
+            page = request.args.get('__page')
+            limit = request.args.get('__limit')
+            books_status = books_status.limit(limit).offset(int(limit) * int(page) - 1)
         if not books_status:
             return make_response(jsonify({'error': 'Resource not found'}), 404)
-        data = self.schema().dump(books_status, many=True)
+        data = self.schema().dump(books_status.all(), many=True)
         return make_response(jsonify(data), 200)
 
+    @jwt_required
     def post(self):
         data = request.json
         if data['status'] == 'borrowed' and data['user_id'] and data['book_id']:
@@ -129,29 +163,21 @@ class BooksStatusResource(Resource):
             db.session.commit()
             return 'Book Status Added Successfully',201
 
-
+    @jwt_required
     def patch(self, slug):
         data = request.json
         obj = BooksStatus.query.get(slug)
-        data = request.json
         if data['status'] == 'free':
             self.schema().load(data, session=db.session, instance=obj)
             obj.book_id = None
             obj.user_id = None
+            obj.borrow_date = None
+            obj.borrow_till = None
 
             db.session.commit()
             db.session.delete(obj)
             db.session.commit()
             return 'User Updated Successfully',200
-
-
-    def delete(self, slug):
-        obj = self.model.query.get(slug)
-        if not obj:
-            return make_response(jsonify({'error': 'Resource not found'}), 404)
-        db.session.delete(obj)
-        db.session.commit()
-        return make_response(jsonify({}), 204)
 
 
 api.add_resource(BooksStatusResource, '/book_status/<slug>', endpoint='book_status')
